Move TerminalConfigHelper, electron-browser->browser
Part of #69115
