+++ src/tsconfig.strictNullChecks.json
@@ -279,6 +279,7 @@
 		"./vs/workbench/contrib/surveys/electron-browser/nps.contribution.ts",
 
 		"./vs/workbench/contrib/terminal/browser/terminalActions.ts",
+		"./vs/workbench/contrib/terminal/browser/terminalConfigHelper.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalFindWidget.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalLinkHandler.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalPanel.ts",
@@ -291,7 +292,6 @@
 		"./vs/workbench/contrib/terminal/common/terminalEnvironment.ts",
 		"./vs/workbench/contrib/terminal/common/terminalMenu.ts",
 		"./vs/workbench/contrib/terminal/common/terminalService.ts",
-		"./vs/workbench/contrib/terminal/electron-browser/terminalConfigHelper.ts",
 		"./vs/workbench/contrib/terminal/electron-browser/terminalInstance.ts",
 		"./vs/workbench/contrib/terminal/electron-browser/terminalProcessManager.ts",
 		"./vs/workbench/contrib/terminal/node/terminal.ts",
