+++ src/vs/workbench/contrib/terminal/common/terminal.ts
@@ -688,3 +688,9 @@ export interface ITerminalProcessExtHostRequest {
 	cols: number;
 	rows: number;
 }
+
+export enum LinuxDistro {
+	Fedora,
+	Ubuntu,
+	Unknown
+}
