+++ src/vs/workbench/contrib/terminal/test/electron-browser/terminalConfigHelper.test.ts
@@ -4,10 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as assert from 'assert';
-import { TerminalConfigHelper } from 'vs/workbench/contrib/terminal/electron-browser/terminalConfigHelper';
+import { TerminalConfigHelper } from 'vs/workbench/contrib/terminal/browser/terminalConfigHelper';
 import { EDITOR_FONT_DEFAULTS } from 'vs/editor/common/config/editorOptions';
-import { isFedora, isUbuntu } from 'vs/workbench/contrib/terminal/node/terminal';
 import { TestConfigurationService } from 'vs/platform/configuration/test/common/testConfigurationService';
+import { LinuxDistro } from 'vs/workbench/contrib/terminal/common/terminal';
 
 suite('Workbench - TerminalConfigHelper', () => {
 	let fixture: HTMLElement;
@@ -21,22 +21,24 @@ suite('Workbench - TerminalConfigHelper', () => {
 		configurationService.setUserConfiguration('editor', { fontFamily: 'foo' });
 		configurationService.setUserConfiguration('terminal', { integrated: { fontFamily: 'bar' } });
 
-		let configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		let configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
 		assert.equal(configHelper.getFont().fontFamily, 'bar', 'terminal.integrated.fontFamily should be selected over editor.fontFamily');
 
 		configurationService.setUserConfiguration('terminal', { integrated: { fontFamily: null } });
 
 		// Recreate config helper as onDidChangeConfiguration isn't implemented in TestConfigurationService
-		configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		configHelper = new TerminalConfigHelper(LinuxDistro.Fedora, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
-		if (isFedora) {
-			assert.equal(configHelper.getFont().fontFamily, '\'DejaVu Sans Mono\', monospace', 'Fedora should have its font overridden when terminal.integrated.fontFamily not set');
-		} else if (isUbuntu) {
-			assert.equal(configHelper.getFont().fontFamily, '\'Ubuntu Mono\', monospace', 'Ubuntu should have its font overridden when terminal.integrated.fontFamily not set');
-		} else {
-			assert.equal(configHelper.getFont().fontFamily, 'foo', 'editor.fontFamily should be the fallback when terminal.integrated.fontFamily not set');
-		}
+		assert.equal(configHelper.getFont().fontFamily, '\'DejaVu Sans Mono\', monospace', 'Fedora should have its font overridden when terminal.integrated.fontFamily not set');
+
+		configHelper = new TerminalConfigHelper(LinuxDistro.Ubuntu, configurationService, null!, null!, null!);
+		configHelper.panelContainer = fixture;
+		assert.equal(configHelper.getFont().fontFamily, '\'Ubuntu Mono\', monospace', 'Ubuntu should have its font overridden when terminal.integrated.fontFamily not set');
+
+		configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
+		configHelper.panelContainer = fixture;
+		assert.equal(configHelper.getFont().fontFamily, 'foo', 'editor.fontFamily should be the fallback when terminal.integrated.fontFamily not set');
 	});
 
 	test('TerminalConfigHelper - getFont fontSize', function () {
@@ -52,7 +54,7 @@ suite('Workbench - TerminalConfigHelper', () => {
 				fontSize: 10
 			}
 		});
-		let configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		let configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
 		assert.equal(configHelper.getFont().fontSize, 10, 'terminal.integrated.fontSize should be selected over editor.fontSize');
 
@@ -65,13 +67,14 @@ suite('Workbench - TerminalConfigHelper', () => {
 				fontSize: 0
 			}
 		});
-		configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		configHelper = new TerminalConfigHelper(LinuxDistro.Ubuntu, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
-		if (isUbuntu) {
-			assert.equal(configHelper.getFont().fontSize, 8, 'The minimum terminal font size (with adjustment) should be used when terminal.integrated.fontSize less than it');
-		} else {
-			assert.equal(configHelper.getFont().fontSize, 6, 'The minimum terminal font size should be used when terminal.integrated.fontSize less than it');
-		}
+		assert.equal(configHelper.getFont().fontSize, 8, 'The minimum terminal font size (with adjustment) should be used when terminal.integrated.fontSize less than it');
+
+		configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
+		configHelper.panelContainer = fixture;
+		assert.equal(configHelper.getFont().fontSize, 6, 'The minimum terminal font size should be used when terminal.integrated.fontSize less than it');
+
 		configurationService.setUserConfiguration('editor', {
 			fontFamily: 'foo'
 		});
@@ -81,7 +84,7 @@ suite('Workbench - TerminalConfigHelper', () => {
 				fontSize: 1500
 			}
 		});
-		configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
 		assert.equal(configHelper.getFont().fontSize, 25, 'The maximum terminal font size should be used when terminal.integrated.fontSize more than it');
 
@@ -94,13 +97,13 @@ suite('Workbench - TerminalConfigHelper', () => {
 				fontSize: null
 			}
 		});
-		configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		configHelper = new TerminalConfigHelper(LinuxDistro.Ubuntu, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
-		if (isUbuntu) {
-			assert.equal(configHelper.getFont().fontSize, EDITOR_FONT_DEFAULTS.fontSize + 2, 'The default editor font size (with adjustment) should be used when terminal.integrated.fontSize is not set');
-		} else {
-			assert.equal(configHelper.getFont().fontSize, EDITOR_FONT_DEFAULTS.fontSize, 'The default editor font size should be used when terminal.integrated.fontSize is not set');
-		}
+		assert.equal(configHelper.getFont().fontSize, EDITOR_FONT_DEFAULTS.fontSize + 2, 'The default editor font size (with adjustment) should be used when terminal.integrated.fontSize is not set');
+
+		configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
+		configHelper.panelContainer = fixture;
+		assert.equal(configHelper.getFont().fontSize, EDITOR_FONT_DEFAULTS.fontSize, 'The default editor font size should be used when terminal.integrated.fontSize is not set');
 	});
 
 	test('TerminalConfigHelper - getFont lineHeight', function () {
@@ -116,7 +119,7 @@ suite('Workbench - TerminalConfigHelper', () => {
 				lineHeight: 2
 			}
 		});
-		let configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		let configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
 		assert.equal(configHelper.getFont().lineHeight, 2, 'terminal.integrated.lineHeight should be selected over editor.lineHeight');
 
@@ -130,7 +133,7 @@ suite('Workbench - TerminalConfigHelper', () => {
 				lineHeight: 0
 			}
 		});
-		configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
 		assert.equal(configHelper.getFont().lineHeight, 1, 'editor.lineHeight should be 1 when terminal.integrated.lineHeight not set');
 	});
@@ -143,7 +146,7 @@ suite('Workbench - TerminalConfigHelper', () => {
 			}
 		});
 
-		let configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		let configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
 		assert.equal(configHelper.configFontIsMonospace(), true, 'monospace is monospaced');
 	});
@@ -155,7 +158,7 @@ suite('Workbench - TerminalConfigHelper', () => {
 				fontFamily: 'sans-serif'
 			}
 		});
-		let configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		let configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
 		assert.equal(configHelper.configFontIsMonospace(), false, 'sans-serif is not monospaced');
 	});
@@ -167,7 +170,7 @@ suite('Workbench - TerminalConfigHelper', () => {
 				fontFamily: 'serif'
 			}
 		});
-		let configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		let configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
 		assert.equal(configHelper.configFontIsMonospace(), false, 'serif is not monospaced');
 	});
@@ -183,7 +186,7 @@ suite('Workbench - TerminalConfigHelper', () => {
 			}
 		});
 
-		let configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		let configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
 		assert.equal(configHelper.configFontIsMonospace(), true, 'monospace is monospaced');
 	});
@@ -199,7 +202,7 @@ suite('Workbench - TerminalConfigHelper', () => {
 			}
 		});
 
-		let configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		let configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
 		assert.equal(configHelper.configFontIsMonospace(), false, 'sans-serif is not monospaced');
 	});
@@ -215,7 +218,7 @@ suite('Workbench - TerminalConfigHelper', () => {
 			}
 		});
 
-		let configHelper = new TerminalConfigHelper(configurationService, null!, null!, null!);
+		let configHelper = new TerminalConfigHelper(LinuxDistro.Unknown, configurationService, null!, null!, null!);
 		configHelper.panelContainer = fixture;
 		assert.equal(configHelper.configFontIsMonospace(), false, 'serif is not monospaced');
 	});
